+++ cli/pom.xml
@@ -5,12 +5,23 @@
     <artifactId>dumpling-parent</artifactId>
     <version>1.1-SNAPSHOT</version>
   </parent>
-  <artifactId>dumpling</artifactId>
+  <artifactId>dumpling-cli</artifactId>
   <packaging>jar</packaging>
-  <name>Dumpling DSL</name>
-  <description>Dumpling core module</description>
+  <name>Dumpling CLI</name>
+  <description>Dumpling command line interface</description>
 
   <dependencies>
+    <dependency>
+      <groupId>com.github.olivergondza</groupId>
+      <artifactId>dumpling</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>com.github.olivergondza</groupId>
+      <artifactId>dumpling-groovy-api</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+
     <dependency>
       <groupId>javax.annotation</groupId>
       <artifactId>jsr250-api</artifactId>
@@ -30,14 +41,6 @@
     <dependency>
       <groupId>org.codehaus.groovy</groupId>
       <artifactId>groovy-groovysh</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>com.sun</groupId>
-      <artifactId>tools</artifactId>
-      <version>1.6</version>
-      <scope>system</scope>
-      <systemPath>${java.home}/../lib/tools.jar</systemPath>
     </dependency>
   </dependencies>
 </project>
