+++ cli/src/main/java/com/github/olivergondza/dumpling/cli/CliCommand.java
