+++ cli/src/main/java/com/github/olivergondza/dumpling/cli/CommandFailedException.java
