+++ cli/src/test/java/com/github/olivergondza/dumpling/cli/GroovyCommandTest.java
