+++ core/src/main/java/com/github/olivergondza/dumpling/query/Deadlocks.java
@@ -33,11 +33,6 @@ import java.util.Set;
 
 import javax.annotation.Nonnull;
 
-import org.kohsuke.args4j.CmdLineException;
-import org.kohsuke.args4j.Option;
-
-import com.github.olivergondza.dumpling.cli.CliCommand;
-import com.github.olivergondza.dumpling.cli.ProcessStream;
 import com.github.olivergondza.dumpling.model.ProcessRuntime;
 import com.github.olivergondza.dumpling.model.ProcessThread;
 import com.github.olivergondza.dumpling.model.ThreadLock;
@@ -67,33 +62,6 @@ public final class Deadlocks implements SingleThreadSetQuery<Deadlocks.Result<?,
             ThreadType extends ProcessThread<ThreadType, SetType, RuntimeType>
     > Result<SetType, RuntimeType, ThreadType> query(@Nonnull SetType threads) {
         return new Result<SetType, RuntimeType, ThreadType>(threads, showStackTraces);
-    }
-
-    public final static class Command implements CliCommand {
-
-        @Option(name = "-i", aliases = {"--in"}, required = true, usage = "Input for process runtime")
-        private ProcessRuntime<?, ?, ?> runtime;
-
-        @Option(name = "--show-stack-traces", usage = "List stack traces of all threads involved")
-        private boolean showStackTraces = false;
-
-        @Override
-        public String getName() {
-            return "deadlocks";
-        }
-
-        @Override
-        public String getDescription() {
-            return "Detect cycles of blocked threads";
-        }
-
-        @Override
-        public int run(@Nonnull ProcessStream process) throws CmdLineException {
-            @SuppressWarnings({"rawtypes", "unchecked"})
-            Result<?, ?, ?> result = new Result(runtime.getThreads(), showStackTraces);
-            result.printInto(process.out());
-            return result.exitCode();
-        }
     }
 
     /**
